requirements.txt
gradio==5.12.0
markdown==3.7
google-cloud-aiplatform==1.79.0
google-auth==2.38.0
firebase-admin==6.6.0
google-cloud-storage==2.19.0
markitdown==0.0.1a3
